Moved to_camel_case, discover_models to ve.core.base_model
